refactor: decompose Import into modular Header/import/ folder structure
